requirements.txt
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.61.1
greenlet==3.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
mysql-connector-python==9.5.0
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.45
tk==0.1.0
typing_extensions==4.15.0
tzdata==2025.3
